🌱 [scanner] refactor: split DashboardState.ts into layout and persistence modules (#21385)
Extract card-layout and drag-collision logic to layout.ts and
storage/cache constants + helpers to persistence.ts, reducing
DashboardState.ts from 993 LOC to ~912 LOC and separating concerns
per the spec in issue #21385.

- layout.ts: dashboardCollisionDetection strategy, POINTER_SENSOR_ACTIVATION_DISTANCE
- persistence.ts: CachedDashboard/PendingDeploy types, dashboardCache
  singleton + setters, AUTO_REFRESH_INTERVAL_MS, DASHBOARD_STORAGE_KEY,
  DEFAULT_DASHBOARD_CARDS, initLocalCardsState()
- DashboardState.ts: imports from new modules; no behaviour changes
